scripts/requirements.txt
supabase==2.3.4
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3 